+++ crates/tad-gui/src/sample_editor.rs
@@ -456,7 +456,7 @@ impl TestSampleWidget {
                     id,
                     PlaySampleArgs {
                         note,
-                        note_length: self.note_length.value() as u32,
+                        note_length: self.note_length.value() as u16,
                         envelope,
                     },
                 ))
